perf: extract shared block-colours utility and consolidate stripExternalLinks
- Create lib/block-colours.ts with blockColours array and getColourFromTitle /
  getColoursFromTitle helpers, removing the same constant defined separately
  in post-header, post-preview, homepage-block, and section-separator
- Export stripExternalLinks from lib/sanitize, removing the identical function
  duplicated in hero-post and post-preview
- Simplify post-body to use lib/sanitize instead of importing DOMPurify directly,
  removing a manual SSR guard already handled by the shared utility
