+++ components/section-separator.tsx
@@ -1,16 +1,7 @@
 import styled from '@emotion/styled';
 import { useEffect, useState } from 'react';
+import { blockColours } from '../lib/block-colours';
 import { colours } from '../pages/_app';
-
-const blockColours = [
-  colours.pink,
-  colours.green,
-  colours.purple,
-  colours.burgandy,
-  colours.dark,
-  colours.azure,
-  colours.blueish,
-];
 
 export default function SectionSeparator() {
   const [backgroundColour, setBackgroundColour] = useState(blockColours[0]);
